Move normalizeUser to user repository
Refactored normalizeUser utility from src/lib/user.ts to src/data/repositories/user/index.ts for better cohesion. Updated all imports and removed the old file. This change centralizes user normalization logic within the user repository.
